backend/requirements.txt
# FinGuide Backend Dependencies
# ==============================

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0,<5.0.0

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# HTTP Client (for external APIs)
httpx>=0.26.0

# Environment Variables
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
black>=24.1.0
isort>=5.13.0
